backend/pyproject.toml
[project]
name = "docgate"
version = "1.1.0"
description = "Guard your docs with supertokens+jwt."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "aiosqlite>=0.22.1",
  "asyncpg>=0.31.0",
  "fastapi[standard-no-fastapi-cloud-cli]>=0.128.0",
  "fs-pyutils>=1.1.1",
  "gunicorn>=25.1.0",
  "pyjwt>=2.10.1",
  "python-dotenv>=1.2.1",
  "sniffio>=1.3.1",
  "sqlalchemy[asyncio]>=2.0.45",
  "stripe>=14.3.0",
  "supertokens-python>=0.30.2",
  "yarl>=1.22.0",
]

[tool.ruff]
line-length = 120
indent-width = 2
exclude = ["scripts", "alembic"]

[tool.ruff.lint]
# 确保 "I" (isort) 规则在启用列表中
select = ["E", "F", "B", "UP", "I", "RET"]

[tool.ruff.lint.isort]
# 这里可以像 isort 一样自定义配置
force-single-line = false # 是否强制单行导入
combine-as-imports = true # 是否合并 import as

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[tool.uv.sources]
fs-pyutils = { git = "https://github.com/fseasy/fs-pyutils.git" }

[tool.ruff.lint.per-file-ignores]
"**/routes*.py" = ["B008"] # avoid warning on `Depends()` in route signature

[tool.mypy]
strict = true
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
files = ["src/docgate"]
exclude = []


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/docgate"]


[dependency-groups]
dev = [
  "alembic>=1.18.3",
  "mypy>=1.19.1",
  "psycopg2-binary>=2.9.11",
  "pytest>=9.0.2",
  "pytest-asyncio>=1.3.0",
  "ruff>=0.14.13",
]
